🏗️ refactor: reorganize repository implementations following VSA+DDD principles
- Move all DrizzleRepository implementations to infrastructure/drizzle directories
- Follow VSA pattern by keeping feature-specific infrastructure within features
- Update .gitignore to allow tracking drizzle directories in features
- Remove repository exports from domain index files
- Rename auth repository to DrizzleAuthUserRepository to avoid confusion

This change aligns with VSA+DDD best practices where:
- Domain layer remains pure with only interfaces
- Infrastructure implementations are colocated with their features
- Clear separation between domain contracts and technical implementations
