+++ apps/api/src/features/quiz/domain/index.ts
@@ -3,9 +3,6 @@
  * @fileoverview Main entry point for quiz domain layer
  */
 
-// Base infrastructure
-export type { Clock } from '@api/shared/clock';
-export { SystemClock } from '@api/shared/clock';
 // Errors
 export { QuizDomainError } from '../shared/errors';
 // Aggregates
@@ -14,7 +11,6 @@ export { AggregateRoot } from './base/AggregateRoot';
 // Entities
 export { Answer } from './entities/Answer';
 export { DomainEvent } from './events/DomainEvent';
-export { DrizzleQuizRepository } from './repositories/DrizzleQuizRepository';
 // Repositories
 export type { IQuizRepository } from './repositories/IQuizRepository';
 export type { Category, Difficulty, ExamType } from './value-objects/ExamTypes';
