+++ apps/api/src/features/user/domain/index.ts
@@ -5,8 +5,6 @@
 
 // Export User entity as User for user bounded context
 export { User } from './entities/User';
-// Repositories
-export { DrizzleUserRepository } from './repositories/DrizzleUserRepository';
 // Export domain types for cross-boundary communication (if needed)
 export type { IUserRepository } from './repositories/IUserRepository';
 // Export value objects for user bounded context
